Move handlePushError from ingester.go to errors.go
