+++ pkg/ingester/errors_test.go
@@ -3,6 +3,7 @@
 package ingester
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"net/http"
@@ -20,6 +21,7 @@ import (
 
 	"github.com/grafana/mimir/pkg/mimirpb"
 	"github.com/grafana/mimir/pkg/util/globalerror"
+	"github.com/grafana/mimir/pkg/util/log"
 	"github.com/grafana/mimir/pkg/util/validation"
 )
 
@@ -312,5 +314,153 @@ func checkIngesterError(t *testing.T, err error, expectedType ingesterErrorType,
 	if isSoft {
 		var softErr softError
 		require.ErrorAs(t, err, &softErr)
+	}
+}
+
+func TestHandlePushError(t *testing.T) {
+	originalMsg := "this is an error"
+	originalErr := errors.New(originalMsg)
+	labelAdapters := []mimirpb.LabelAdapter{{Name: labels.MetricName, Value: "testmetric"}, {Name: "foo", Value: "biz"}}
+	labels := mimirpb.FromLabelAdaptersToLabels(labelAdapters)
+
+	timestamp := model.Time(1)
+	testCases := []struct {
+		name                string
+		err                 error
+		doNotLogExpected    bool
+		expectedTranslation error
+	}{
+		{
+			name:                "a generic error is not translated",
+			err:                 originalErr,
+			expectedTranslation: originalErr,
+		},
+		{
+			name:                "a DoNotLog error of a generic error is not translated",
+			err:                 log.DoNotLogError{Err: originalErr},
+			expectedTranslation: log.DoNotLogError{Err: originalErr},
+			doNotLogExpected:    true,
+		},
+		{
+			name:                "an unavailableError gets translated into an errorWithStatus Unavailable error",
+			err:                 newUnavailableError(services.Stopping),
+			expectedTranslation: newErrorWithStatus(newUnavailableError(services.Stopping), codes.Unavailable),
+		},
+		{
+			name: "a wrapped unavailableError gets translated into a non-loggable errorWithStatus Unavailable error",
+			err:  fmt.Errorf("wrapped: %w", newUnavailableError(services.Stopping)),
+			expectedTranslation: newErrorWithStatus(
+				fmt.Errorf("wrapped: %w", newUnavailableError(services.Stopping)),
+				codes.Unavailable,
+			),
+		},
+		{
+			name: "an instanceLimitReachedError gets translated into a non-loggable errorWithStatus Unavailable error",
+			err:  newInstanceLimitReachedError("instance limit reached"),
+			expectedTranslation: newErrorWithStatus(
+				log.DoNotLogError{Err: newInstanceLimitReachedError("instance limit reached")},
+				codes.Unavailable,
+			),
+			doNotLogExpected: true,
+		},
+		{
+			name: "a wrapped instanceLimitReachedError gets translated into a non-loggable errorWithStatus Unavailable error",
+			err:  fmt.Errorf("wrapped: %w", newInstanceLimitReachedError("instance limit reached")),
+			expectedTranslation: newErrorWithStatus(
+				log.DoNotLogError{Err: fmt.Errorf("wrapped: %w", newInstanceLimitReachedError("instance limit reached"))},
+				codes.Unavailable,
+			),
+			doNotLogExpected: true,
+		},
+		{
+			name: "a tsdbUnavailableError gets translated into an errorWithHTTPStatus 503 error",
+			err:  newTSDBUnavailableError("tsdb stopping"),
+			expectedTranslation: newErrorWithHTTPStatus(
+				newTSDBUnavailableError("tsdb stopping"),
+				http.StatusServiceUnavailable,
+			),
+		},
+		{
+			name: "a wrapped tsdbUnavailableError gets translated into an errorWithHTTPStatus 503 error",
+			err:  fmt.Errorf("wrapped: %w", newTSDBUnavailableError("tsdb stopping")),
+			expectedTranslation: newErrorWithHTTPStatus(
+				fmt.Errorf("wrapped: %w", newTSDBUnavailableError("tsdb stopping")),
+				http.StatusServiceUnavailable,
+			),
+		},
+		{
+			name: "a sampleError gets translated into an errorWithHTTPStatus 400 error",
+			err:  newSampleError("id", "sample error", timestamp, labelAdapters),
+			expectedTranslation: newErrorWithHTTPStatus(
+				newSampleError("id", "sample error", timestamp, labelAdapters),
+				http.StatusBadRequest,
+			),
+		},
+		{
+			name: "a wrapped exemplarError gets translated into an errorWithHTTPStatus 400 error",
+			err:  fmt.Errorf("wrapped: %w", newSampleError("id", "sample error", timestamp, labelAdapters)),
+			expectedTranslation: newErrorWithHTTPStatus(
+				fmt.Errorf("wrapped: %w", newSampleError("id", "sample error", timestamp, labelAdapters)),
+				http.StatusBadRequest,
+			),
+		},
+		{
+			name: "a exemplarError gets translated into an errorWithHTTPStatus 400 error",
+			err:  newExemplarError("id", "exemplar error", timestamp, labelAdapters, labelAdapters),
+			expectedTranslation: newErrorWithHTTPStatus(
+				newExemplarError("id", "exemplar error", timestamp, labelAdapters, labelAdapters),
+				http.StatusBadRequest,
+			),
+		},
+		{
+			name: "a wrapped exemplarError gets translated into an errorWithHTTPStatus 400 error",
+			err:  fmt.Errorf("wrapped: %w", newExemplarError("id", "exemplar error", timestamp, labelAdapters, labelAdapters)),
+			expectedTranslation: newErrorWithHTTPStatus(
+				fmt.Errorf("wrapped: %w", newExemplarError("id", "exemplar error", timestamp, labelAdapters, labelAdapters)),
+				http.StatusBadRequest,
+			),
+		},
+		{
+			name: "a perUserMetadataLimitReachedError gets translated into an errorWithHTTPStatus 400 error",
+			err:  newPerUserSeriesLimitReachedError(10),
+			expectedTranslation: newErrorWithHTTPStatus(
+				newPerUserSeriesLimitReachedError(10),
+				http.StatusBadRequest,
+			),
+		},
+		{
+			name: "a wrapped perUserMetadataLimitReachedError gets translated into an errorWithHTTPStatus 400 error",
+			err:  fmt.Errorf("wrapped: %w", newPerUserSeriesLimitReachedError(10)),
+			expectedTranslation: newErrorWithHTTPStatus(
+				fmt.Errorf("wrapped: %w", newPerUserSeriesLimitReachedError(10)),
+				http.StatusBadRequest,
+			),
+		},
+		{
+			name: "a perMetricMetadataLimitReachedError gets translated into an errorWithHTTPStatus 400 error",
+			err:  newPerMetricSeriesLimitReachedError(10, labels),
+			expectedTranslation: newErrorWithHTTPStatus(
+				newPerMetricSeriesLimitReachedError(10, labels),
+				http.StatusBadRequest,
+			),
+		},
+		{
+			name: "a wrapped perMetricMetadataLimitReachedError gets translated into an errorWithHTTPStatus 400 error",
+			err:  fmt.Errorf("wrapped: %w", newPerMetricSeriesLimitReachedError(10, labels)),
+			expectedTranslation: newErrorWithHTTPStatus(
+				fmt.Errorf("wrapped: %w", newPerMetricSeriesLimitReachedError(10, labels)),
+				http.StatusBadRequest,
+			),
+		},
+	}
+
+	for _, tc := range testCases {
+		handledErr := handlePushError(tc.err)
+		require.Equal(t, tc.expectedTranslation, handledErr)
+		if tc.doNotLogExpected {
+			var doNotLogError log.DoNotLogError
+			require.ErrorAs(t, handledErr, &doNotLogError)
+			require.False(t, doNotLogError.ShouldLog(context.Background(), 0))
+		}
 	}
 }
